feat: Support mutations
Ref: #51
