+++ src/hypothesis_graphql/_strategies/context.py
@@ -1,0 +1,9 @@
+import attr
+import graphql
+
+
+@attr.s(slots=True)
+class Context:
+    """The common context for query generation."""
+
+    schema: graphql.GraphQLSchema = attr.ib()
