+++ src/hypothesis_graphql/strategies.py
@@ -1,3 +1,4 @@
 # pylint: disable=unused-import
+from ._strategies.mutations import mutations
 from ._strategies.queries import query
 from ._strategies.schema import schema
