+++ test/conftest.py
@@ -8,6 +8,74 @@ settings.register_profile("default", suppress_health_check=[HealthCheck.too_slow
 settings.load_profile("default")
 
 
+SCHEMA = """
+type Book {
+  title: String
+  author: Author
+}
+
+type Author {
+  name: String
+  books: [Book]
+}
+
+enum Color {
+  RED
+  GREEN
+  BLUE
+}
+
+input EnumInput {
+  color: Color
+}
+
+input QueryInput {
+  eq: String
+  ne: String
+}
+
+input RequiredInput {
+  eq: Float!,
+}
+
+input NestedQueryInput {
+  code: QueryInput
+}
+
+type Image {
+  path: String
+}
+type Video {
+  duration: Int
+}
+
+union Media = Image | Video
+
+interface Node {
+  id: ID
+}
+
+interface Alone {
+  id: ID
+}
+
+type Model implements Node {
+  int: Int,
+  float: Float,
+  media: Media,
+  string: String
+  id: ID,
+  boolean: Boolean
+  color: Color
+}
+"""
+
+
+@pytest.fixture(scope="session")
+def schema():
+    return SCHEMA
+
+
 @pytest.fixture(scope="session")
 def build_schema():
     # Parses a GraphQL schema. Caching is required to avoid re-parsing in each Hypothesis test as schemas are mostly
@@ -21,7 +89,7 @@ def build_schema():
 
 
 @pytest.fixture(scope="session")
-def validate_query(build_schema):
+def validate_operation(build_schema):
     def inner(schema, query):
         if isinstance(schema, str):
             parsed_schema = build_schema(schema)
