+++ test/test_queries.py
@@ -5,75 +5,13 @@ from hypothesis import assume, find, given
 from hypothesis import strategies as st
 
 from hypothesis_graphql import strategies as gql_st
-from hypothesis_graphql._strategies.queries import value_nodes
-
-SCHEMA = """
-type Book {
-  title: String
-  author: Author
-}
-
-type Author {
-  name: String
-  books: [Book]
-}
-
-enum Color {
-  RED
-  GREEN
-  BLUE
-}
-
-input EnumInput {
-  color: Color
-}
-
-input QueryInput {
-  eq: String
-  ne: String
-}
-
-input RequiredInput {
-  eq: Float!,
-}
-
-input NestedQueryInput {
-  code: QueryInput
-}
-
-type Image {
-  path: String
-}
-type Video {
-  duration: Int
-}
-
-union Media = Image | Video
-
-interface Node {
-  id: ID
-}
-
-interface Alone {
-  id: ID
-}
-
-type Model implements Node {
-  int: Int,
-  float: Float,
-  media: Media,
-  string: String
-  id: ID,
-  boolean: Boolean
-  color: Color
-}
-"""
+from hypothesis_graphql._strategies.selections import value_nodes
 
 
 @pytest.fixture(scope="session")
-def simple_schema():
+def simple_schema(schema):
     return (
-        SCHEMA
+        schema
         + """type Query {
           getBooks: [Book]
           getAuthors: [Author]
@@ -94,16 +32,16 @@ def simple_schema():
     ),
 )
 @given(data=st.data())
-def test_query(data, query_type, validate_query):
-    schema = SCHEMA + query_type
+def test_query(data, schema, query_type, validate_operation):
+    schema = schema + query_type
     query = data.draw(gql_st.query(schema))
-    validate_query(schema, query)
+    validate_operation(schema, query)
 
 
 @given(data=st.data())
-def test_query_subset(data, simple_schema, validate_query):
+def test_query_subset(data, simple_schema, validate_operation):
     query = data.draw(gql_st.query(simple_schema, fields=["getBooks"]))
-    validate_query(simple_schema, query)
+    validate_operation(simple_schema, query)
     assert "getAuthors" not in query
 
 
@@ -118,13 +56,13 @@ def test_wrong_fields(simple_schema):
 
 
 @given(data=st.data())
-def test_query_from_graphql_schema(data, validate_query):
+def test_query_from_graphql_schema(data, schema, validate_operation):
     query = """type Query {
       getBooksByAuthor(name: String): [Book]
     }"""
-    schema = graphql.build_schema(SCHEMA + query)
+    schema = graphql.build_schema(schema + query)
     query = data.draw(gql_st.query(schema))
-    validate_query(schema, query)
+    validate_operation(schema, query)
 
 
 @pytest.mark.parametrize("notnull", (True, False))
@@ -156,16 +94,16 @@ def test_query_from_graphql_schema(data, validate_query):
     ),
 )
 @given(data=st.data())
-def test_arguments(data, arguments, node_names, notnull, validate_query):
+def test_arguments(data, schema, arguments, node_names, notnull, validate_operation):
     if notnull:
         arguments += "!"
     query_type = f"""type Query {{
       getModel({arguments}): Model
     }}"""
 
-    schema = SCHEMA + query_type
+    schema = schema + query_type
     query = data.draw(gql_st.query(schema))
-    validate_query(schema, query)
+    validate_operation(schema, query)
     for node_name in node_names:
         assert node_name not in query
     if notnull:
@@ -187,11 +125,11 @@ def test_arguments(data, arguments, node_names, notnull, validate_query):
     ),
 )
 @given(data=st.data())
-def test_interface(data, query_type, validate_query):
-    schema = SCHEMA + query_type
+def test_interface(data, schema, query_type, validate_operation):
+    schema = schema + query_type
     parsed_schema = graphql.build_schema(schema)
     query = data.draw(gql_st.query(schema))
-    validate_query(parsed_schema, query)
+    validate_operation(parsed_schema, query)
 
 
 @pytest.mark.parametrize(
@@ -219,8 +157,8 @@ def test_interface(data, query_type, validate_query):
         ),
     ),
 )
-def test_minimal_queries(query, minimum):
-    schema = SCHEMA + f"type Query {{ {query} }}"
+def test_minimal_queries(query, schema, minimum):
+    schema = schema + f"type Query {{ {query} }}"
     strategy = gql_st.query(schema)
     minimal_query = f"""{{
   getAuthors{minimum} {{
@@ -261,18 +199,18 @@ type Query {{
 
 
 @given(data=st.data())
-def test_custom_scalar_non_argument(data, validate_query):
+def test_custom_scalar_non_argument(data, validate_operation):
     # When a custom scalar type is defined
     # And is used in a non-argument position
 
     schema = CUSTOM_SCALAR_TEMPLATE.format(query="getObjects: [Object]")
     query = data.draw(gql_st.query(schema))
-    validate_query(schema, query)
+    validate_operation(schema, query)
     # Then queries should be generated
     assert "created" in query
 
 
-def test_custom_scalar_argument_nullable(validate_query):
+def test_custom_scalar_argument_nullable(validate_operation):
     # When a custom scalar type is defined
     # And is used in an argument position
     # And is nullable
@@ -287,7 +225,7 @@ def test_custom_scalar_argument_nullable(validate_query):
         nonlocal num_of_queries
 
         num_of_queries += 1
-        validate_query(schema, query)
+        validate_operation(schema, query)
         assert "getByDate {" in query
 
     test()
@@ -308,7 +246,7 @@ def test_custom_scalar_argument(data):
 
 
 @given(data=st.data())
-def test_no_surrogates(data, validate_query):
+def test_no_surrogates(data, validate_operation):
     # Unicode surrogates are not supported by GraphQL spec
     schema = """
     type Query {
@@ -316,7 +254,7 @@ def test_no_surrogates(data, validate_query):
     }
     """
     query = data.draw(gql_st.query(schema))
-    document = validate_query(schema, query)
+    document = validate_operation(schema, query)
     argument_node = document.definitions[0].selection_set.selections[0].arguments[0]
     assume(argument_node.name.value == "user")
     value = argument_node.value.value
